+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Get OKAPI version and tenant module info after login. Refs STRIPES-671.
 * Mock okapi session in local storage for testing. Refs STCOR-444.
 * Handle `react-router-dom` deprecation warnings. Refs STCOR-448.
+* Update `react-intl` to `v5`. Refs STCOR-449.
 
 ## [5.0.2](https://github.com/folio-org/stripes-core/tree/v5.0.2) (2020-06-12)
 [Full Changelog](https://github.com/folio-org/stripes-core/compare/v5.0.1...v5.0.2)